Renamed codegen library to navigationreactnative
Autolinking on Android doesn't like hyphens in library name
