+++ NavigationReactNative/src/android/build.gradle
@@ -68,7 +68,7 @@ dependencies {
 if (isNewArchitectureEnabled()) {
     react {
         jsRootDir = file("../")
-        libraryName = "navigation-react-native"
+        libraryName = "navigationreactnative"
         codegenJavaPackageName = "com.navigation.reactnative"
     }
 }
