+++ NavigationReactNative/src/cpp/react/renderer/components/navigation-react-native/NVActionBarShadowNode.h
@@ -1,8 +1,8 @@
 #pragma once
 
 #include "NVActionBarState.h"
-#include <react/renderer/components/navigation-react-native/EventEmitters.h>
-#include <react/renderer/components/navigation-react-native/Props.h>
+#include <react/renderer/components/navigationreactnative/EventEmitters.h>
+#include <react/renderer/components/navigationreactnative/Props.h>
 #include <react/renderer/components/view/ConcreteViewShadowNode.h>
 
 namespace facebook {
