+++ NavigationReactNative/src/cpp/react/renderer/components/navigation-react-native/NVBarButtonShadowNode.h
@@ -2,8 +2,8 @@
 
 #include "NVBarButtonState.h"
 #include <react/renderer/components/view/ConcreteViewShadowNode.h>
-#include <react/renderer/components/navigation-react-native/EventEmitters.h>
-#include <react/renderer/components/navigation-react-native/Props.h>
+#include <react/renderer/components/navigationreactnative/EventEmitters.h>
+#include <react/renderer/components/navigationreactnative/Props.h>
 #include <react/renderer/imagemanager/ImageManager.h>
 
 namespace facebook {
