+++ NavigationReactNative/src/ios/NVActionBarComponentView.mm
@@ -1,11 +1,11 @@
 #ifdef RCT_NEW_ARCH_ENABLED
 #import "NVActionBarComponentView.h"
 
-#import <react/renderer/components/navigation-react-native/ComponentDescriptors.h>
-#import <react/renderer/components/navigation-react-native/EventEmitters.h>
-#import <react/renderer/components/navigation-react-native/Props.h>
-#import <react/renderer/components/navigation-react-native/RCTComponentViewHelpers.h>
-#import <navigation-react-native/NVActionBarComponentDescriptor.h>
+#import <react/renderer/components/navigationreactnative/ComponentDescriptors.h>
+#import <react/renderer/components/navigationreactnative/EventEmitters.h>
+#import <react/renderer/components/navigationreactnative/Props.h>
+#import <react/renderer/components/navigationreactnative/RCTComponentViewHelpers.h>
+#import <navigationreactnative/NVActionBarComponentDescriptor.h>
 
 #import "RCTFabricComponentsPlugins.h"
