+++ NavigationReactNative/src/ios/NVExtendedFloatingActionButtonComponentView.mm
@@ -1,10 +1,10 @@
 #ifdef RCT_NEW_ARCH_ENABLED
 #import "NVExtendedFloatingActionButtonComponentView.h"
 
-#import <react/renderer/components/navigation-react-native/ComponentDescriptors.h>
-#import <react/renderer/components/navigation-react-native/EventEmitters.h>
-#import <react/renderer/components/navigation-react-native/Props.h>
-#import <react/renderer/components/navigation-react-native/RCTComponentViewHelpers.h>
+#import <react/renderer/components/navigationreactnative/ComponentDescriptors.h>
+#import <react/renderer/components/navigationreactnative/EventEmitters.h>
+#import <react/renderer/components/navigationreactnative/Props.h>
+#import <react/renderer/components/navigationreactnative/RCTComponentViewHelpers.h>
 
 #import "RCTFabricComponentsPlugins.h"
